update benchmark files
