+++ benchmarks/dask/env.yml
@@ -73,10 +73,10 @@ dependencies:
   - lcms2=2.16=ha0e7c42_0
   - lerc=4.0.0=h9a09cb3_0
   - libabseil=20240722.0=cxx17_hf9b8971_1
-  - libarrow=18.1.0=h654e1bb_2_cpu
-  - libarrow-acero=18.1.0=h605b82c_2_cpu
-  - libarrow-dataset=18.1.0=h605b82c_2_cpu
-  - libarrow-substrait=18.1.0=h9b432b6_2_cpu
+  - libarrow=16.1.0=h654e1bb_46_cpu
+  - libarrow-acero=16.1.0=h605b82c_46_cpu
+  - libarrow-dataset=16.1.0=h605b82c_46_cpu
+  - libarrow-substrait=16.1.0=h9b432b6_46_cpu
   - libblas=3.9.0=25_osxarm64_openblas
   - libbrotlicommon=1.1.0=hd74edd7_2
   - libbrotlidec=1.1.0=hd74edd7_2
@@ -100,7 +100,7 @@ dependencies:
   - liblapack=3.9.0=25_osxarm64_openblas
   - libnghttp2=1.64.0=h6d7220d_0
   - libopenblas=0.3.28=openmp_hf332438_1
-  - libparquet=18.1.0=h5168bdf_2_cpu
+  - libparquet=16.1.0=h5168bdf_46_cpu
   - libpng=1.6.44=hc14010f_0
   - libprotobuf=5.28.2=h8f0b736_0
   - libre2-11=2024.07.02=h2348fd5_1
@@ -143,8 +143,8 @@ dependencies:
   - pthread-stubs=0.4=hd74edd7_1002
   - ptyprocess=0.7.0=pyhd3deb0d_0
   - pure_eval=0.2.3=pyhd8ed1ab_0
-  - pyarrow=18.1.0=py310hb6292c7_0
-  - pyarrow-core=18.1.0=py310hc17921c_0_cpu
+  - pyarrow=16.1.0=py310h24597f5_6
+  - pyarrow-core=16.1.0=py310hc17921c_6_cpu
   - pycparser=2.22=pyh29332c3_1
   - pygments=2.18.0=pyhd8ed1ab_1
   - pysocks=1.7.1=pyha55dd90_7
